requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
prometheus-client==0.16.0
structlog==23.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
asyncpg==0.27.0
httpx==0.24.1
locust==2.16.0
